refactor(appointment): 패키지 정리
