+++ backend/src/main/java/com/morak/back/appointment/application/AppointmentService.java
@@ -3,9 +3,9 @@ package com.morak.back.appointment.application;
 import com.morak.back.appointment.domain.Appointment;
 import com.morak.back.appointment.domain.AppointmentRepository;
 import com.morak.back.appointment.domain.AvailableTime;
-import com.morak.back.appointment.domain.RankRecommendation;
+import com.morak.back.appointment.domain.recommend.RankRecommendation;
 import com.morak.back.appointment.domain.SystemTime;
-import com.morak.back.appointment.domain.RecommendationCells;
+import com.morak.back.appointment.domain.recommend.RecommendationCells;
 import com.morak.back.appointment.exception.AppointmentAuthorizationException;
 import com.morak.back.appointment.exception.AppointmentNotFoundException;
 import com.morak.back.appointment.ui.dto.AppointmentAllResponse;
