+++ backend/src/main/java/com/morak/back/appointment/domain/Appointment.java
@@ -2,6 +2,11 @@ package com.morak.back.appointment.domain;
 
 import static com.morak.back.core.exception.CustomErrorCode.AVAILABLETIME_OUT_OF_RANGE_ERROR;
 
+import com.morak.back.appointment.domain.dateperiod.DatePeriod;
+import com.morak.back.appointment.domain.menu.Menu;
+import com.morak.back.appointment.domain.menu.MenuStatus;
+import com.morak.back.appointment.domain.recommend.AppointmentTime;
+import com.morak.back.appointment.domain.timeperiod.TimePeriod;
 import com.morak.back.appointment.exception.AppointmentAuthorizationException;
 import com.morak.back.appointment.exception.AppointmentDomainLogicException;
 import com.morak.back.auth.domain.Member;
