+++ backend/src/main/java/com/morak/back/appointment/domain/recommend/RecommendationCells.java
@@ -1,5 +1,7 @@
-package com.morak.back.appointment.domain;
+package com.morak.back.appointment.domain.recommend;
 
+import com.morak.back.appointment.domain.Appointment;
+import com.morak.back.appointment.domain.AvailableTime;
 import com.morak.back.auth.domain.Member;
 import java.util.ArrayList;
 import java.util.List;
