+++ backend/src/main/java/com/morak/back/appointment/ui/dto/AppointmentStatusResponse.java
@@ -1,6 +1,6 @@
 package com.morak.back.appointment.ui.dto;
 
-import com.morak.back.appointment.domain.MenuStatus;
+import com.morak.back.appointment.domain.menu.MenuStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
