+++ backend/src/main/java/com/morak/back/appointment/ui/dto/RecommendationResponse.java
@@ -1,7 +1,7 @@
 package com.morak.back.appointment.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.morak.back.appointment.domain.RankRecommendation;
+import com.morak.back.appointment.domain.recommend.RankRecommendation;
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.ui.dto.MemberResponse;
 import java.time.LocalDateTime;
