+++ backend/src/test/java/com/morak/back/appointment/domain/AppointmentTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.morak.back.appointment.domain.Appointment.AppointmentBuilder;
+import com.morak.back.appointment.domain.menu.MenuStatus;
+import com.morak.back.appointment.domain.recommend.AppointmentTime;
 import com.morak.back.appointment.exception.AppointmentAuthorizationException;
 import com.morak.back.appointment.exception.AppointmentDomainLogicException;
 import com.morak.back.auth.domain.Member;
