+++ backend/src/test/java/com/morak/back/appointment/domain/MenuStatusTest.java
@@ -3,6 +3,7 @@ package com.morak.back.appointment.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.morak.back.appointment.domain.menu.MenuStatus;
 import com.morak.back.appointment.exception.AppointmentDomainLogicException;
 import com.morak.back.core.exception.CustomErrorCode;
 import org.junit.jupiter.api.Test;
